feature(bento-pan-zoom): renamed hook file
